refactor(sdk): Move `find_event_relations` and `find_event_with_relations` to `persistence`.
This patch moves the `find_event_relations` and
`find_event_with_relations` implementations in `persistence` with the
idea to share it with the thread caches.
